finally rework
