+++ BooksStore/WebClient/ClientApp/src/app/body/home/advertising/adversting.component.html
@@ -1,0 +1,37 @@
+<div fxFlex fxLayout="row" class="main-carousel">
+
+<div fxFlex class="imageSlider">
+    <button class="prevSlide"></button>
+    <button class="nextSlide"></button>
+    <div class="imageimageSlider_container">
+
+      <div class="imageSlider_slide slideActive">
+        <img
+          src="https://images.unsplash.com/photo-1568143100438-58f5b47e0a9a?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1351&q=80"
+          alt="" style=" border-radius: 5px;">
+      </div>
+
+      <div class="imageSlider_slide">
+        <a href="#"><img
+            src="https://images.unsplash.com/photo-1568164528001-fdf61ebe1a7e?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1351&q=80"
+            alt="" style=" border-radius: 5px;"></a>
+      </div>
+
+      <div class="imageSlider_slide">
+        <a href="#"><img
+            src="https://images.unsplash.com/photo-1568186704829-7daaf3327194?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1350&q=80"
+            alt="" style=" border-radius: 5px;"></a>
+      </div>
+    </div>
+  </div>
+
+  <div fxLayout="column" style="background: white; border-radius: 5px;">
+    <div style=" margin: 5px;"> <img
+        src="https://images.unsplash.com/photo-1568143100438-58f5b47e0a9a?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1351&q=80"
+        alt="" style="height:200px; width:400px; margin: 0px;  border-radius: 5px;"></div>
+    <div style=" margin: 5px;">
+      <img
+        src="https://images.unsplash.com/photo-1568143100438-58f5b47e0a9a?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1351&q=80"
+        alt="" style="height:200px; width:400px; margin: 0px; border-radius: 5px; "></div>
+  </div>
+  </div>
