+++ BooksStore/WebClient/ClientApp/src/app/body/home/bestsellers/bestsellers.component.css
@@ -1,0 +1,6 @@
+.main-list {
+  background: white; 
+  margin: 5px; 
+  border-radius: 5px; 
+  box-shadow: 0 3px 5px rgba(0, 0, 0, .6);
+}
