+++ BooksStore/WebClient/ClientApp/src/app/body/home/bestsellers/bestsellers.component.html
@@ -1,0 +1,148 @@
+<div>
+    <div fxLayout="column" class="main-list">
+      <div fxLayout="row">
+        <mat-card class="min-card">
+          <mat-card-title>
+            Simple card
+          </mat-card-title>
+          <mat-card-content>
+            <img
+              src='https://ndc.book24.ru/resize_cache/9658024/82aff9b7004fb1831f626fc857f53daf/iblock/332/332600fc6d027a75a5f2e99a18d5069d.jpg'
+              alt='image' class="img-min-card">
+          </mat-card-content>
+        </mat-card>
+        <mat-card class="min-card">
+          <mat-card-title>
+            Simple card
+          </mat-card-title>
+          <img
+            src='https://ndc.book24.ru/resize_cache/9658024/82aff9b7004fb1831f626fc857f53daf/iblock/332/332600fc6d027a75a5f2e99a18d5069d.jpg'
+            alt='image' class="img-min-card">
+        </mat-card>
+        <mat-card class="min-card">
+          <mat-card-title>
+            Simple card
+          </mat-card-title>
+          <img src='https://cdn.book24.ru/v2/ITD000000000994236/COVER/cover3d1__w130.webp' alt='image'
+            class="img-min-card">
+        </mat-card>
+        <mat-card class="min-card">
+          <mat-card-title>
+            Simple card
+          </mat-card-title><img src='https://cdn.book24.ru/v2/ASE000000000844497/COVER/cover3d1__w130.webp' alt='image'
+            class="img-min-card">
+        </mat-card>
+        <mat-card class="min-card">
+          <mat-card-title>
+            Simple card
+          </mat-card-title><img src='https://cdn.book24.ru/v2/AST000000000155650/COVER/cover3d1__w130.webp' alt='image'
+            class="img-min-card">
+        </mat-card>
+        <mat-card class="min-card">
+          <mat-card-title>
+            Simple card
+          </mat-card-title><img src='https://cdn.book24.ru/v2/AST000000000155650/COVER/cover3d1__w130.webp' alt='image'
+            class="img-min-card">
+        </mat-card>
+        <mat-card class="min-card">
+          <mat-card-title>
+            Simple card
+          </mat-card-title><img src='https://cdn.book24.ru/v2/AST000000000155650/COVER/cover3d1__w130.webp' alt='image'
+            class="img-min-card">
+        </mat-card>
+      </div>
+      <div fxLayout="row">
+        <mat-card class="min-card">
+          <mat-card-content>
+            <mat-card-title>
+              Simple card
+            </mat-card-title>
+            <img src='https://cdn.book24.ru/v2/ITD000000000995229/COVER/cover3d1__w130.webp' alt='image'
+              class="img-min-card">
+          </mat-card-content>
+        </mat-card>
+        <mat-card class="min-card">
+          <mat-card-title>
+            Simple card
+          </mat-card-title><img src='https://cdn.book24.ru/v2/AST000000000130417/COVER/cover3d1__w130.webp' alt='image'
+            class="img-min-card">
+        </mat-card>
+        <mat-card class="min-card">
+          <mat-card-title>
+            Simple card
+          </mat-card-title><img src='https://cdn.book24.ru/v2/ITD000000000985056/COVER/cover3d1__w130.webp' alt='image'
+            class="img-min-card">
+        </mat-card>
+        <mat-card class="min-card">
+          <mat-card-title>
+            Simple card
+          </mat-card-title><img src='https://cdn.book24.ru/v2/ASE000000000844497/COVER/cover3d1__w130.webp' alt='image'
+            class="img-min-card">
+        </mat-card>
+        <mat-card class="min-card">
+          <mat-card-title>
+            Simple card
+          </mat-card-title><img src='https://cdn.book24.ru/v2/AST000000000155650/COVER/cover3d1__w130.webp' alt='image'
+            class="img-min-card">
+        </mat-card>
+        <mat-card class="min-card">
+          <mat-card-title>
+            Simple card
+          </mat-card-title><img src='https://cdn.book24.ru/v2/AST000000000155650/COVER/cover3d1__w130.webp' alt='image'
+            class="img-min-card">
+        </mat-card>
+        <mat-card class="min-card">
+          <mat-card-title>
+            Simple card
+          </mat-card-title><img src='https://cdn.book24.ru/v2/AST000000000155650/COVER/cover3d1__w130.webp' alt='image'
+            class="img-min-card">
+        </mat-card>
+      </div>
+      <div fxLayout="row">
+        <mat-card class="min-card">
+          <mat-card-content>
+            <mat-card-title>
+              Simple card
+            </mat-card-title><img src='https://cdn.book24.ru/v2/ASE000000000844497/COVER/cover3d1__w130.webp'
+              class="img-min-card">
+          </mat-card-content>
+        </mat-card>
+        <mat-card class="min-card">
+          <mat-card-title>
+            Simple card
+          </mat-card-title><img src='https://cdn.book24.ru/v2/ASE000000000844497/COVER/cover3d1__w130.webp' alt='image'
+            class="img-min-card">
+        </mat-card>
+        <mat-card class="min-card">
+          <mat-card-title>
+            Simple card
+          </mat-card-title><img src='https://cdn.book24.ru/v2/AST000000000155650/COVER/cover3d1__w130.webp' alt='image'
+            class="img-min-card">
+        </mat-card>
+        <mat-card class="min-card">
+          <mat-card-title>
+            Simple card
+          </mat-card-title><img src='https://cdn.book24.ru/v2/ASE000000000844497/COVER/cover3d1__w130.webp' alt='image'
+            class="img-min-card">
+        </mat-card>
+        <mat-card class="min-card">
+          <mat-card-title>
+            Simple card
+          </mat-card-title><img src='https://cdn.book24.ru/v2/AST000000000155650/COVER/cover3d1__w130.webp' alt='image'
+            class="img-min-card">
+        </mat-card>
+        <mat-card class="min-card">
+          <mat-card-title>
+            Simple card
+          </mat-card-title><img src='https://cdn.book24.ru/v2/AST000000000155650/COVER/cover3d1__w130.webp' alt='image'
+            class="img-min-card">
+        </mat-card>
+        <mat-card class="min-card">
+          <mat-card-title>
+            Simple card
+          </mat-card-title><img src='https://cdn.book24.ru/v2/AST000000000155650/COVER/cover3d1__w130.webp' alt='image'
+            class="img-min-card">
+        </mat-card>
+      </div>
+    </div>
+  </div>
