+++ BooksStore/WebClient/ClientApp/src/app/body/home/bestsellers/bestsellers.component.ts
@@ -1,0 +1,16 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-bestsellers',
+  templateUrl: './bestsellers.component.html',
+  styleUrls: ['./bestsellers.component.css']
+})
+
+
+
+export class BestsellersComponent implements OnInit {
+  constructor() {
+  }
+  ngOnInit() {
+  }
+}
