--- BooksStore/WebClient/ClientApp/src/app/body/home/home.component.css
@@ -1,150 +1,0 @@
-.component {
-  position: relative;
-  width: 80%;
-  display: block;
-  margin: 0 auto;
-  padding: 0;
-  text-align: center;
-}
-
-.main-carousel {
-  background: white; 
-  margin: 5px 5px 15px 5px; 
-  border-radius: 5px;  
-  box-shadow: 0 3px 5px rgba(0, 0, 0, .6);
-}
-
-.main-list {
-  background: white; 
-  margin: 5px; 
-  border-radius: 5px; 
-  box-shadow: 0 3px 5px rgba(0, 0, 0, .6);
-}
-
-.imageSlider {
-  height: 430px;
-  overflow: hidden;
-  background: rgba(199, 201, 211, 0.15);
-  position: relative;
-  width: calc(63%);
-  display: inline-block;
-  margin: 15px 10px 0px 10px;
-}
-
-  .imageSlider button {
-    display: block;
-    width: 40px;
-    height: 100%;
-    position: absolute;
-    top: 0;
-    bottom: 0;
-    border: 0;
-    background: rgba(0, 0, 0, 0);
-    z-index: 3;
-    outline: none;
-    cursor: pointer;
-  }
-
-    .imageSlider button.prevSlide {
-      left: 0;
-      background: url('data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAADIAAAAyCAQAAAC0NkA6AAAAAmJLR0QA/4ePzL8AAAB3SURBVFjD7dgxCoAwDIXhntCqKB307kIFFXqECr+bdHEzmd67wAfJEPJCUHzDwElhsiQSFYBiT8BlT1RGG2J+iZtFhAgRIkSI+CD65iQlq7u3mxNeiMe4QmBqFr+KESNGjBgxDo9p9GCyR1lw2NYekY1MpwLopzwbyhH0gn4t6QAAAABJRU5ErkJggg==') no-repeat center rgba(0, 0, 0, 0);
-      background-size: 24px;
-    }
-
-    .imageSlider button.nextSlide {
-      right: 0;
-      background: url('data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAADIAAAAyCAQAAAC0NkA6AAAAAmJLR0QA/4ePzL8AAAB3SURBVFjD7dgxCoAwEETReMIEUVLo2S0SFERvEOHbSRo7syDMXODBbrHsOKd8E0ZOVkJLouMAoBBbMjs0Z+gpFszwMBezGDFixIgR04iJ1VkLFky2QNJ/x8VYLX4SIUKECBEiXghv8JiSLX7fzaIs8CQWvAog49znkhL0iBSaKAAAAABJRU5ErkJggg==') no-repeat center rgba(0, 0, 0, 0);
-      background-size: 24px;
-    }
-
-    .imageSlider button:hover {
-      background-color: rgba(0, 0, 0, 0.15);
-    }
-
-.imageSlider .imageimageSlider_container {
-  height: 100%;
-  position: relative;
-}
-
-.imageSlider .imageimageSlider_container .imageSlider_slide {
-  position: absolute;
-  clear: both;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  width: 100%;
-  height: 100%;
-  margin: 0 auto;
-  overflow: hidden;
-  opacity: 0;
-}
-
-.imageSlider .imageimageSlider_container .imageSlider_slide.slideActive {
-  z-index: 2;
-  opacity: 1;
-  transition: 1s;
-}
-
-.imageSlider .imageimageSlider_container .imageSlider_slide a {
-  display: block;
-  text-decoration: none;
-  width: 100%;
-  height: 100%;
-}
-
-.imageSlider .imageimageSlider_container .imageSlider_slide img {
-  display: block;
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  object-position: center;
-  vertical-align: top;
-}
-
-.imageSlider .imageimageSlider_container .imageSlider_slide span {
-  display: block;
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  z-index: 1;
-  background: rgba(255, 255, 255, 0.65);
-  line-height: 1.35;
-  padding: 6px 15px;
-  font-size: 12px;
-}
-
-.imageSlider .rightDivContent {
-  margin: 15px auto;
-  border-radius: 3px;
-}
-
-#container {
-  width: 99%;
-  margin: 10px 10px 10px 10px;
-  padding: 0;
-  background: white;
-  border-radius: 5px;
-  box-shadow: 0 3px 5px rgba(0, 0, 0, .6);
-}
-
-.items {
-  max-width: 300px;
-  height: 300px;
-}
-
-#left, #right {
-  /* margin:30px; */
-  height: 20%;
-  width: 20%;
-}
-
-::ng-deep .ng2-carouselamos-wrapper {
-  width: 90% !important;
-}
-
-::ng-deep #right.ng-star-inserted, #left.ng-star-inserted {
-  width: 30px !important;
-  height: 30px !important;
-}
