+++ BooksStore/WebClient/ClientApp/src/app/body/home/popular/popular.component.css
@@ -1,0 +1,28 @@
+#container {
+  width: 99%;
+  margin: 10px 10px 10px 10px;
+  padding: 0;
+  background: white;
+  border-radius: 5px;
+  box-shadow: 0 3px 5px rgba(0, 0, 0, .6);
+}
+
+.items {
+  max-width: 300px;
+  height: 300px;
+}
+
+#left, #right {
+  /* margin:30px; */
+  height: 20%;
+  width: 20%;
+}
+
+::ng-deep .ng2-carouselamos-wrapper {
+  width: 90% !important;
+}
+
+::ng-deep #right.ng-star-inserted, #left.ng-star-inserted {
+  width: 30px !important;
+  height: 30px !important;
+}
