+++ BooksStore/WebClient/ClientApp/src/app/body/home/popular/popular.component.html
@@ -1,0 +1,25 @@
+<div id="container">
+    Популярные категории
+    <div ng2-carouselamos class="slides-wrapper" [items]="items" [width]="900" [$prev]="prev" [$next]="next"
+      [$item]="item">
+    </div>
+    <ng-template style="width: 20px" #prev>
+      <img
+        src='https://c7.uihere.com/files/204/55/746/creative-commons-license-non-commercial-wikimedia-commons-public-domain-angle-thumb.jpg'
+        style="transform: scaleX(-1);" id="left">
+    </ng-template>
+    <ng-template style="width: 20px" #next>
+      <img
+        src='https://c7.uihere.com/files/204/55/746/creative-commons-license-non-commercial-wikimedia-commons-public-domain-angle-thumb.jpg'
+        style="height:20%; width: 20%;" id="right">
+    </ng-template>
+    <ng-template #item let-item let-i="index">
+      <div class="items">
+        <mat-card style="height: 300px; width:200px; margin:30px">
+          <mat-card-title>
+            {{item.book}}
+          </mat-card-title><img src='{{item.name}}' class="img-min-card" alt='image'>
+        </mat-card>
+      </div>
+    </ng-template>
+  </div>
